requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backoff==1.11.1
base45==0.4.3
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.50.0
django-environ==0.8.1
frozenlist==1.3.0
idna==3.3
ipify==1.0.0
ipinfo==4.2.0
ipinfo-django==1.2.0
multidict==6.0.2
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
